Rename CurveBackend to AlgDncBackend (its Curves + Algebra)
And update git ref to func tests and mpa ext
